Changes the asimov lawset so its less prone to abuse
You may not injure any being instead of 
You may not injure a human being .

This fixes the problems with asimov where it let the borgs or AI kill people as it fucking wanted because they were a lizard. Now everyone is protected as it was originally planned to be until some speshul snowflakes added other races.