[WIP] Adds Text-to-Speech (#8584)
* Adds TTS

* Rewrite a lot of TTS subsystem

Owner is now a client instead of a mob. Voice is stored against DNA so lings can inherit it when they transform into another person. Doing some basic sanitisation on the message to stop radio messages from starting with "semicolon", etc. Added better cooldown handling

* Fixed DNA and speech modifiers

* Optimised code, moved generator to tools folder

* Added new preference, reorganised code

* Added missing comment, removed redundant file

* Removed redundant file

* Reworked the processing method

* Fixed paths

* Refactors the say_tts proc to be more like playsound

Turns out using a custom channel was actually a bad idea

* Removed useless channel

* Reset game_options.txt

* Fixed typo in preferences

* Added HUD indicator

* Clarified channel var name

* Reworked how TTS generator creates files. Hopefull we won't be prematurely sending files

* Changed up say() proc, we now run after say() and use our own sanitation

* Added back speech mods

But there is a problem we will have to deal with soon

* Added volume modifier

* Fixed borgs not having TTS

* Updated tts_generator

* Minor fixes

* Completely revamped TTS processing

Now the engine is open all the time and requests are sent via files so no chance of someone hijacking shell()

* Fixed lock check

* Deleted log file

whoops!

* Reworked process further

* Adds debug button to start engine