[MIRROR] Clown Car balance changes (#9074)
* Clown Car balance changes (#40222)

balance: The Clown Car can no longer move unrestricted in zero gravity environments.
balance: The Clown Car now costs 20 TC to purchase.

This PR was done at the request of @vuonojenmustaturska. I renamed entered.dm to sealed.dm because it's where the sealed subtype is defined and it's primarily used for procs specific to sealed. The balance changes were Naksu's idea. I definitely think not being able to move freely in space is a good change. I'm ambivalent about the TC, but the item felt like a gimmick that you should fully commit to like His Grace, and the free space movement + 4 TC space suit was a bit cancerous.

* Clown Car balance changes

* Update uplink_items.dm

* Update hippiestation.dme

* Update uplink_items.dm

* Update uplink_items.dm