Capping volume pumps done even better
This was tested on homeserver and is working. As long as the maxcap is set for a pressure is set, the pump will never go beyond it. As this is a pretty reliable way to do so, I thought why not be bold enough to increase the amount of max pressure while staying non-explosive