[READY]Further nerfing atmos stuff (#9088)
* Further nerfing atmos stuff

I'll make columbine look like a fucking joke

* Update atmospherics.dm

* getting the right undefines

* Update atmospherics.dm

* [Unfinished]: Welcome to quality assessment and control

* Courtesy of one of our codebuse testers

No name will be revealed to keep anonymity. If this all nerfs bombs into nonexistence, we could maybe talk about increasing transfer efficiency back to normal

* Update atmospherics.dm

* Update atmospherics.dm

* ah, so that parenthesis messes up the whole compilation

good to know

* no need for initiating datums

must have been the alcohol again

* Update atmospherics.dm

* yoyo's review

* removing the new list initiation on each tick

* fixed a slight error

* sometimes, getting drunk helps you mistakes you wouldn't notice otherwise

1 oz lemon juice, 2.5 oz amaretto, filled up to 11-12 oz strawberry juice, to be put in a blender filled with ice, then roughly sieved into creamy goodness

* drop everything, I got an idea

* Capping volume pumps done even better

This was tested on homeserver and is working. As long as the maxcap is set for a pressure is set, the pump will never go beyond it. As this is a pretty reliable way to do so, I thought why not be bold enough to increase the amount of max pressure while staying non-explosive

* minor push ref fix

* Update atmospherics.dm

* I just had a talk with Naksu. He told me:

!var does not equal var <= 0. Whoever told me this: You're wrong.

* fixing a possible runtime

* Update atmospherics.dm