Alien infestation event not limited to once per round, but will not fire while existing aliens are alive
:cl: coiax
tweak: Alien infestations are not limited to once per round only, but will never
happen while existing aliens are alive.
/:cl:

Essentially, it's really disheartening when a larva is spawned, and it dies quickly
because the player makes a mistake or is unlucky. So, aliens MAY happen again, but
it's still a weight 5 event, so don't get your hopes up. PR: https://github.com/tgstation/tgstation/pull/41719