Fix extraneous naked human icon appearing below all humans (#42155)
How this is happening (to my knowledge) is this:

Although 81Denton was fixing a broken icon_state what he failed to realize
(I'm not the only one who doesn't test okay?) is that we use overlays to render
all parts of a human, rather than a single sprite. Hell the sprite it was
changed to doesn't even have directionals. That sprite was displayed instead of
the usually "broken" non-existing icon (it didn't exist because we never used
it due to the fact we render every part of a human including bodyparts with
overlays instead of a base icon, after dismemberment I believe?) Yeah. This
fixes that by nulling the icon_state so it's humans are invisible by default
before their overlays are applied (which is the behavior before Denton's PR).