Adds Blank Paper, a new mime cocktail (#42366)
Adds a new drink, made by mimes, for everyone! It was based in a costum drink that had the flavour text of eraser, nothing and future!

Recipe is 1 part Silencer, 1 part Nothing and 1 Part Nuka Cola (it also heals mime like most mime drinks do)

cl BebeYoshi & Hexmaniacosanna
add: A new drink called Blank Paper was added to the bar menu, it was made by a mime and it represents a new start.
/cl

[why]: This was a special costum drink and we thought it deserved its own cocktail, besides, mimes needs more unique things (even if you don't even need a mime to make it).