fixes chilling infinite timestop (#42432)
cl no
fix: Chilling Sepia extracts cannot be used infinitely and will be consumed upon use.
/cl

[why]:

Chilling sepia extracts, unlike all other crossbred chilling extracts, were not consumed after activation with plasma and thus could be spammed constantly if you injected them with plasma and then activated them, with absolutely no delay or cooldown in any form.

This is probably a bug, given that cerulean extracts, the ACTUAL infinite use extracts, have a delay on how many times you can activate them in a quick timespan and will stop working if you activate it within the cooldown, and that the description text itself describes the extract shattering.

Even if it wasn't a bug, infinite timestop with no cooldown and superior to the actual infinite extract is bad and needs to go.