Adds wooden chapel pews (#42712)
* Adds wooden chapel pews

This adds wooden chapel pews to the game. They function much like ratty couches, in that there are several different pieces which can be mixed and matched to make different length pews. They can be built from 3 wooden sheets.

The change to sheet_types.dm adds the pew crafting recipies to the end of the wood sheets crafting list.

The added file pew.dm  is the code which makes the pews work.

The added file pew.dmi is the sprites for the pews, including the overlays for the armrests.

the change to tgstation.dme is a single line which includes pew.dm.

This is my first PR ever, so please let me know if I made any mistakes in approved process.

* merged pew.dmi into sofa.dmi

As per the instructions of several commenters, I merged pew.dmi into sofa.dmi. pew.dm has been updated to reflect this.