Crossbow Fixes
Makes a crossbow cut all previous overlays before applying another set of them whenever its update_icon() proc is called.

When the examine proc is called, a crossbow now checks that there is a chambered projectile rod instead of trying to look at the BB variable in order to avoid a runtime error that comes from checking the BB variable from null, and the text that is shown when this check is passed is now changed so that it will not show the loaded rod in the form "A the metal rod", but rather "A metal rod".