Fix high security door material exploit, tweaks cost (#43035)
Fixes an exploit that allowed for infinite metal/plasteel to be farmed off of
high security/vault door deconstruction. Makes completed airlock assemblies
start with no wire security and adjusts the material costs of both doors
accordingly. As the high security airlock was fortified with normal metal
unlike the vault door this is a slight buff to it, but I am fairly sure it
will not affect the game balance noticeably.