[READY] Adds the Gravitokinetic stand type (#43846)
This adds the gravitokinetic stand. It's attacks are weaker but they apply heavy gravity.
They can also alt click the ground to apply heavy gravity to that tile, the gravity applied from floor punches is heavy enough to damage people over time.

Recalling drops all of the affected tiles, and they have a range as well.