Fixes mixing beer with other drinks breaking the sprite (#44145)
About The Pull Request

fixes: #43070
All credit for this should go to @ExcessiveUseOfCobblestone as he made a comment on the issue that fixed it perfectly. All I did was implement it since it wasn't in the code yet and the bug was still there.
Why It's Good For The Game

Bug fixes are always good.
Changelog

cl Fire Chance
fix: Fixed the drink sprites overlapping when you are mixing beer with other drinks.
/cl