Fixes diagonal box movement, makes stealth box reveal when damaged (#45274)
About The Pull Request

Lowers the integrity of the implant box from 70 (normal cardboard boxes) to 1
fixes #38612
Why It's Good For The Game

The stealth box is already an extremely powerful item, mostly due to a bug #38612 that has been around ever since we could move diagonally and hasn't been fixed.

The box is supposed to be comparable to the other stealthy item, the Chameleon Projector. However, the box has several large advantages, including:
-Full invisibility
-Faster movement (especially with the bug)
-Less obvious when moving around
-The box takes more hits to reveal you
-Easier to get back to being hard to find when revealed
-You need an open hand to open the box manually, putting you in melee range and without a weapon in the selected hand, leaving you at a major disadvantage.
-The projector can be taken away from you / dropped and takes up inventory space
-The projector needs to be in hand to use, implant is hands free
-Projector needs to be in your in-use hand to disable yourself, the box is a button, meaning you can be more prepared with the box to strike out from stealth.

The projector still has a niche, however the box is currently the meta stealth item to pick.

This PR fixes the movement bug (no more wavedashing) and makes any damage to the box reveal the user, no longer letting it be a massive damage sponge. This should bring the box much closer in line to where the Chameleon Projector is, balance wise.
Changelog

cl Chen Yakumo
balance: The stealth implant box is now flimsier than it was before.
fix: You can no longer zoom around in a box by hugging walls.
/cl