Double Barrel Shotgun Icon State Fixing
Renames icon states in the projectile.dmi file so that all of the possible icon states for a double barrel shotgun are valid.

Makes the '-sawn' suffix be used when a current skin is chosen as well, since it was initially being used by all of the possible icon_states of the sawn-off double barrel shotgun.