Bug Fixes
Things with abductor tongues now don't gibber something they said with the change being that we set the message to empty text instead of returning empty text, since the proc that sends a signal to call the handle_speech proc wouldn't stop the message from being sent if it isn't empty.

The A variable of the overridden after_attack proc for damp rags is now assumed to be an atom instead of an obj, since it would lead the first check to conclude that mobs aren't valid if you only assume it to be an obj, which in turn stops you from using rags to smother mobs.

Seeds using the hippiestation/icons/obj/hydroponics/seeds.dmi file now don't become invisible after being modified with a plant DNA manipulator as now every seed is forced to use the icons/obj/hydroponics/seeds.dmi file in the repaint_seed proc.

MMIs and/or positronic brains now use their parent canUseTopic proc whenever they aren't trying to use an electronic assembly, and when they are using an electronic assembly, they use the check_bot_self variable as the return value, which is True only if they are trying to look at the open assembly they were inserted into.

The overridden canUseTopic proc for pAIs was also rewritten to make it use the parent canUseTopic proc when the pAI doesn't try to access the topic of an electronic assembly or an integrated circuit.