Makes railings good (#49171)
About The Pull Request

Makes rails functional in that they can be vaulted from either side, moved, and repaired. Still doesn't add a way to make them.
Why It's Good For The Game

make multiz great

Waiting/working on a few things for the map so I'm just atomizing the few things I'm sitting on so other map makers can take advantage of in the meantime.