Nuclear Operatives will now see the nuke codes in memories (#61448)
This adds the ability for the nuclear operatives to check their memories for the nuke code. Memories are the successor to notes, the latter being where the nuke code was stored before. Many thanks to tralezab for helping me make this change. Also shoutout MrDoomBringer for making the messages suck less
Conflicts:
	code/__DEFINES/memory_defines.dm
	strings/memories.json